前端用户功能域完成。
git-svn-id: https://apposs-center.googlecode.com/svn/trunk@8 e3cf8a15-5072-d663-4c44-a26c1ac8da71

+++ app/controllers/cmd_set_defs_controller.rb
@@ -1,13 +1,22 @@
 class CmdSetDefsController < BaseController
   def index
     app_id = params[:app_id]
-    respond_with current_app.cmd_set_defs.collect { |obj|
-      obj.serializable_hash.update("actions" => [
-          {:name=>"准备执行", :flex => 1.7, :url=> app_cmd_sets_path(app_id), :type => 'simple', :method => 'POST'},
-          {:name=>"修改", :flex => 1, :url=> edit_app_cmd_set_def_path(app_id, obj.id), :type => 'multi', :method => 'GET'},
-          {:name=>"删除", :flex => 1, :url=> app_cmd_set_def_path(app_id, obj.id), :type => 'delete', :method => 'DELETE'}
-      ],"flex" => 8)
-    }
+    role = Stakeholder.where(:app_id => app_id, :user_id => current_user.id).first.role_id
+    if role == Role::Admin || role == Role::PE
+      respond_with current_app.cmd_set_defs.collect { |obj|
+        obj.serializable_hash.update("actions" => [
+            {:name=>"准备执行", :flex => 1.7, :url=> app_cmd_sets_path(app_id), :type => 'simple', :method => 'POST'},
+            {:name=>"修改", :flex => 1, :url=> edit_app_cmd_set_def_path(app_id, obj.id), :type => 'multi', :method => 'GET'},
+            {:name=>"删除", :flex => 1, :url=> app_cmd_set_def_path(app_id, obj.id), :type => 'delete', :method => 'DELETE'}
+        ], "flex" => 8, "add" => true)
+      }
+    else
+      respond_with current_app.cmd_set_defs.collect { |obj|
+        obj.serializable_hash.update("actions" => [
+            {:name=>"准备执行", :flex => 1, :url=> app_cmd_sets_path(app_id), :type => 'simple', :method => 'POST'}
+        ], "flex" => 5, "add" => false)
+      }
+    end
   end
 
   def show
